+++ Actions/PhotonNetworkGetLastMessagePlayerProperties.cs
@@ -1,4 +1,4 @@
-// (c) Copyright HutongGames, LLC 2010-2012. All rights reserved.
+// (c) Copyright HutongGames, LLC 2010-2015. All rights reserved.
 
 using UnityEngine;
 
@@ -110,7 +110,7 @@ namespace HutongGames.PlayMaker.Actions
 				{
 					return false;
 				}
-				PlayMakerPhotonProxy.ApplyValueToFsmVar(this.Fsm,customPropertiesValues[i],_player.customProperties[key.Value]);
+				PlayMakerUtils.ApplyValueToFsmVar(this.Fsm,customPropertiesValues[i],_player.customProperties[key.Value]);
 				i++;
 			}