+++ Actions/PhotonNetworkGetNextPlayerProperties.cs
@@ -1,4 +1,4 @@
-// (c) Copyright HutongGames, LLC 2010-2013. All rights reserved.
+// (c) Copyright HutongGames, LLC 2010-2015. All rights reserved.
 
 using UnityEngine;
 using System.Collections.Generic;
@@ -140,7 +140,7 @@ namespace HutongGames.PlayMaker.Actions
 			int i = 0;
 			foreach(FsmString key in customPropertyKeys)
 			{
-				PlayMakerPhotonProxy.ApplyValueToFsmVar(this.Fsm,customPropertiesValues[i],_player.customProperties[key.Value]);
+				PlayMakerUtils.ApplyValueToFsmVar(this.Fsm,customPropertiesValues[i],_player.customProperties[key.Value]);
 				i++;
 			}