+++ Actions/PhotonNetworkGetNextRoomProperties.cs
@@ -1,4 +1,4 @@
-// (c) Copyright HutongGames, LLC 2010-2013. All rights reserved.
+// (c) Copyright HutongGames, LLC 2010-2015. All rights reserved.
 
 using UnityEngine;
 using System.Collections.Generic;
@@ -138,7 +138,7 @@ namespace HutongGames.PlayMaker.Actions
 			{
 				if (_room.customProperties.ContainsKey(key.Value))
 				{
-					PlayMakerPhotonProxy.ApplyValueToFsmVar(this.Fsm,customPropertiesValues[i],_room.customProperties[key.Value]);
+					PlayMakerUtils.ApplyValueToFsmVar(this.Fsm,customPropertiesValues[i],_room.customProperties[key.Value]);
 				}
 				i++;
 			}